cctv/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
